fix/Расширил компонент
