+++ src/components/ui/droplist/droplist.tsx
@@ -1,13 +1,12 @@
 import React, { FC, useEffect, useState, /* SetStateAction */ } from 'react';
 import cn from 'classnames';
 
-import { Icon } from '../icon';
-
 import styles from './droplist.module.css';
 
 // Компоненты
 import DroplistItems from './droplist-items/droplist-items';
 import ListSelected from './list-selected/list-selected';
+import ContainerButton from './container-button/container-button';
 
 interface IDroplistProps {
   dataType: Array<string> | Array<number>,
@@ -23,10 +22,6 @@ export const Droplist: FC<IDroplistProps> = ({ dataType, handlerSubmitDroplist }
   // Выбран ли Dropdown
   const [ activeDropdown, setActiveDropdown ] = useState(false);
 
-  const clickActiveDropdown = (): void => {
-    setActiveDropdown(!activeDropdown);
-  };
-
   useEffect(() => {
     getList(dataType);
   }, []);
@@ -38,17 +33,7 @@ export const Droplist: FC<IDroplistProps> = ({ dataType, handlerSubmitDroplist }
 
   return (
     <div className={ cn(styles.dropdown) }>
-      <div
-        className={ cn(styles.container) }
-        onClick={ clickActiveDropdown }
-      >
-        <p className={ cn(styles.text) }>
-          Все
-        </p>
-        {
-          <Icon glyph='arrow-down' fill='black' className={ styles.iconArrowDown } />
-        }
-      </div>
+      <ContainerButton setActiveDropdown={ setActiveDropdown } />
       <form
         name='dropdown'
         className={ cn(styles.form) }
